pytest: Separate proquint tests
